requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pdfplumber
requests
pydantic
python-multipart
PyPDF2>=3.0.0
psycopg2-binary
python-jose[cryptography]
passlib
email-validator
apscheduler
google-cloud-storage
